requirements.txt
Flask
Flask-SocketIO
GitPython
eventlet
streamlit
